python_test/requirements.txt

asgiref==3.5.2
click==8.1.3
colorama==0.4.4
Django==4.0.4
pip==21.2.3
PyMySQL==1.0.2
setuptools==57.4.0
sqlparse==0.4.2
tzdata==2022.1
pathlib2==2.3.6
mysql-connector==2.2.9